Move all the ReadRows functionality over
